Rename schema.js to joi.helpers.js
